+++ CHANGES.md
@@ -5,11 +5,15 @@
 ### Improvements
 
 -   Support writing dataframes without geometry column (#267)
-
 -   Calculate feature count by iterating over features if GDAL returns an
     unknown count for a data layer (e.g., OSM driver); this may have signficant
     performance impacts for some data sources that would otherwise return an
     unknown count (count is used in `read_info`, `read`, `read_dataframe`) (#271).
+-   Raise error if `read` or `read_dataframe` is called with parameters to read no
+    columns, geometry, or fids (#280)
+
+-   Automatically detect supported driver by extension for all available
+    write drivers and addition of `detect_write_driver` (#270)
 
 ### Bug fixes
 
@@ -18,6 +22,8 @@
     `use_arrow=True` (#265)
 -   Fix errors reading OSM data due to invalid feature count and incorrect
     reading of OSM layers beyond the first layer (#271)
+-   Always raise an exception if there is an error when writing a data source
+    (#284)
 
 ## 0.6.0 (2023-04-27)